test: 추가된 메서드들에 대한 테스트코드 작성
