+++ src/test/java/blackjack/domain/participant/DealerTest.java
@@ -1,6 +1,7 @@
 package blackjack.domain.participant;
 
 import blackjack.domain.card.Card;
+import blackjack.domain.card.CardDeck;
 import blackjack.domain.card.CardNumber;
 import blackjack.domain.card.CardType;
 import org.junit.jupiter.api.BeforeEach;
@@ -36,55 +37,11 @@ public class DealerTest {
     }
 
     @Test
-    @DisplayName("딜러가 카드를 받았는지 확인")
-    void receiveCard() {
-        dealer.receiveOneCard(new Card(CardNumber.ACE, CardType.CLOVER));
-        assertThat(dealer.cardCount()).isEqualTo(1);
-    }
-
-    @Test
-    @DisplayName("딜러가 카드를 더 받을 수 있는지 확인")
-    void checkMoreCardAvailable() {
-        dealer.receiveOneCard(new Card(CardNumber.TWO, CardType.CLOVER));
-        dealer.receiveOneCard(new Card(CardNumber.TEN, CardType.CLOVER));
-        assertThat(dealer.checkMoreCardAvailable()).isTrue();
-    }
-
-    @Test
-    @DisplayName("딜러가 갖고 있는 카드의 합을 확인")
-    void calculateCardSum() {
-        dealer.receiveOneCard(new Card(CardNumber.TWO, CardType.CLOVER));
-        dealer.receiveOneCard(new Card(CardNumber.TEN, CardType.CLOVER));
-        assertThat(dealer.calculateScore()).isEqualTo(12);
-    }
-
-    @Test
-    @DisplayName("에이스 카드가 하나있을 때 합 구하기")
-    void calculateCardSumWhenAceIsOne() {
-        dealer.receiveOneCard(new Card(CardNumber.ACE, CardType.CLOVER));
-        dealer.receiveOneCard(new Card(CardNumber.TWO, CardType.CLOVER));
-        assertThat(dealer.calculateScore()).isEqualTo(13);
-    }
-
-    @ParameterizedTest
-    @CsvSource(value = {"ACE,ACE:12", "ACE,ACE,ACE:13", "ACE,ACE,TEN:12"}, delimiter = ':')
-    @DisplayName("에이스 카드가 여러 개일 때 합 구하기")
-    void calculateCardSumWhenAceIsTwo(final String input, final int expected) {
-        final String[] inputs = input.split(",");
-        for (final String number : inputs) {
-            final CardNumber cardNumber = CardNumber.valueOf(number);
-            dealer.receiveOneCard(new Card(cardNumber, CardType.CLOVER));
-        }
-        assertThat(dealer.calculateScore()).isEqualTo(expected);
-    }
-
-    @Test
-    @DisplayName("딜러가 버스트인지 확인")
-    void isBust() {
-        dealer.receiveOneCard(new Card(CardNumber.TEN, CardType.CLOVER));
-        dealer.receiveOneCard(new Card(CardNumber.NINE, CardType.HEART));
-        dealer.receiveOneCard(new Card(CardNumber.EIGHT, CardType.HEART));
-        assertThat(dealer.isBust()).isTrue();
+    @DisplayName("딜러의 초기 카드 출력이 한 장만 되는지 확인")
+    void showInitialCards() {
+        CardDeck cardDeck = new CardDeck();
+        dealer.receiveInitialCards(cardDeck);
+        assertThat(dealer.showInitialCards().size()).isEqualTo(1);
     }
 
     @ParameterizedTest
